# contracts/openapi/mobile-v1.yaml
openapi: 3.1.0
info:
  title: 404 Mobile API
  version: 0.1.0
  description: >
    Modular monolith baseline contract for the 404 mobile application.
servers:
  - url: /api/v1
...
paths:
...
  /facades/notifications:
    get:
      tags: [Facade - Notification]
      summary: Get notification feed payload
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Notification feed payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationFacadeResponse'
...
  /houses:
    post:
      tags: [House]
      summary: Create a house
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseRequest'
      responses:
        '201':
          description: House created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateHouseResponse'
  /houses/join:
    post:
      tags: [House]
      summary: Join house by invite code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinHouseRequest'
      responses:
        '200':
          description: Joined house
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinHouseResponse'
  /houses/me:
    get:
      tags: [House]
      summary: Get active house
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active house summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseSummary'
  /houses/{houseId}/spaces:
    parameters:
      - $ref: '#/components/parameters/HouseId'
    get:
      tags: [Space]
      summary: List spaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Space list
          content:
            application/json:
              schema:
                type: object
                properties:
                  spaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Space'
    post:
      tags: [Space]
      summary: Add space
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpaceRequest'
      responses:
        '201':
          description: Space created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
...
  /houses/{houseId}/notifications:
    parameters:
      - $ref: '#/components/parameters/HouseId'
    get:
      tags: [Notification]
      summary: List house notifications
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationItem'
                  nextCursor:
                    type: string
                    nullable: true
  /notifications/{notificationId}/read:
    parameters:
      - $ref: '#/components/parameters/NotificationId'
    post:
      tags: [Notification]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Marked as read
  /push-devices:
    post:
      tags: [Notification]
      summary: Register push device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPushDeviceRequest'
      responses:
        '201':
          description: Push device registered
  /houses/{houseId}/weekly-recaps/latest:
    parameters:
      - $ref: '#/components/parameters/HouseId'
    get:
      tags: [Review]
      summary: Get latest weekly recap
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Latest weekly recap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyRecapResponse'
  /houses/{houseId}/weekly-recaps/{snapshotId}/satisfaction:
    parameters:
      - $ref: '#/components/parameters/HouseId'
      - $ref: '#/components/parameters/SnapshotId'
    post:
      tags: [Review]
      summary: Submit weekly satisfaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeeklySatisfactionRequest'
      responses:
        '204':
          description: Satisfaction submitted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HouseId:
      in: path
      name: houseId
      required: true
      schema:
        type: string
        format: uuid
...
    NotificationId:
      in: path
      name: notificationId
      required: true
      schema:
        type: string
        format: uuid
    SnapshotId:
      in: path
      name: snapshotId
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    NotificationFacadeResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NotificationItem'
        unreadCount:
          type: integer
        nextCursor:
          type: string
          nullable: true
...
    CreateHouseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        initialCleanlinessVote:
          type: string
          nullable: true
    CreateHouseResponse:
      type: object
      properties:
        houseId:
          type: string
          format: uuid
        membershipId:
          type: string
          format: uuid
        inviteCode:
          $ref: '#/components/schemas/InviteCode'
    JoinHouseRequest:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
    JoinHouseResponse:
      type: object
      properties:
        house:
          $ref: '#/components/schemas/HouseSummary'
        membershipId:
          type: string
          format: uuid
    HouseSummary:
      type: object
      properties:
        houseId:
          type: string
          format: uuid
        name:
          type: string
        cleanlinessLevel:
          type: string
          nullable: true
        memberCount:
          type: integer
    InviteCode:
      type: object
      properties:
        code:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
    CreateSpaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        sortOrder:
          type: integer
          nullable: true
    Space:
      type: object
      properties:
        spaceId:
          type: string
          format: uuid
        name:
          type: string
        sortOrder:
          type: integer
...
    NotificationItem:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        occurredAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    RegisterPushDeviceRequest:
      type: object
      required: [platform, pushToken]
      properties:
        platform:
          type: string
        pushToken:
          type: string
        deviceId:
          type: string
          nullable: true
        appVersion:
          type: string
          nullable: true
    WeeklyRecapResponse:
      type: object
      properties:
        snapshotId:
          type: string
          format: uuid
        weekStartDate:
          type: string
          format: date
        weekEndDate:
          type: string
          format: date
        houseStats:
          type: object
          additionalProperties: true
        memberStats:
          type: array
          items:
            type: object
            additionalProperties: true
        satisfactionSubmitted:
          type: boolean
    WeeklySatisfactionRequest:
      type: object
      required: [score]
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true